{"schema":"decodex/commit/1","summary":"Split MCP planning modules","authority":"XY-1115"}
